requirements.txt
boto3
click
pillow
halo